web/requirements.txt
Flask==2.0.1
Flask-Discord==0.1.64
python-dotenv==0.18.0
pymongo==3.12.0